pyproject.toml
[project]
name = "deepseek-distilled-on-sagemaker"
version = "1.0.0"
description = "A code repository for deploying distilled versions of DeepSeek-R1 on SageMaker (or any other Open Source Model)"
authors = [
    { name = "Bojan Jakimovski - Shekswess", email = "bojan.jakimovski@loka.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.36.16",
    "loguru>=0.7.3",
    "pre-commit>=4.1.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.5",
    "sagemaker>=2.239.0",
]


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 90
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["C901", "E501"]

fixable = ["ALL"]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"

docstring-code-format = true

docstring-code-line-length = "dynamic"
